requirements.txt
GitPython==2.0.2